image_crawling_project/requirements.txt
scrapy
scrapy-splash
cookiecutter
selenium
scrapy-selenium
pymongo
python-dotenv
Pillow
